+++ init.rb
@@ -6,11 +6,11 @@ Redmine::Plugin.register :redmine_dashboard do
   name 'Redmine Dashboard plugin'
   author 'Jan Graichen'
   description 'Add a task board and a planning board to Redmine'
-  version '2.12.4'
+  version '2.13.0'
   url 'https://github.com/jgraichen/redmine_dashboard'
   author_url 'mailto:jgraichen@altimos.de'
 
-  requires_redmine version_or_higher: '4.0'
+  requires_redmine '4.2'
 
   project_module :dashboard do
     permission :view_dashboards, {